Extract the topology provider interface
